Unsubcribe test works
