+++ api/controllers/users.js
@@ -1,5 +1,6 @@
 var mongoose = require("mongoose");
 var User = require("../models/user");
+var Number = require("../models/number");
 const axios = require("axios");
 
 mongoose.model("User");
@@ -33,7 +34,7 @@ exports.getUser = (req, res) => {
   });
 };
 
-exports.addUser = (req, res) => {
+exports.addUser = async function addUser(req, res) {
   console.log(req.body);
   if (
     !req.body.firstName ||
@@ -51,27 +52,37 @@ exports.addUser = (req, res) => {
     });
     return;
   } else {
-    //console.log("BODY: " + req.body.firstName);
-    var newUser = new User({
-      firstName: req.body.firstName,
-      lastName: req.body.lastName,
-      phoneNumber: req.body.phoneNumber,
-      username: req.body.username,
-      subscribed: true,
-      groupID: req.body.groupID
-    });
-    //console.log("User:" + newUser);
-    // attempt to save the user
-    newUser.save(err => {
-      if (err) {
-        console.log(err);
-        return res.json({ success: false, message: err });
-      }
-      res.json({
-        success: true,
-        message: "Successfully created new user"
+    
+    try {
+      //get the next number
+      //should add a check if there is no next number
+      var nreq = await axios.get(`http://localhost:9000/numbers/next`);
+
+      var newUser = new User({
+        firstName: req.body.firstName,
+        lastName: req.body.lastName,
+        phoneNumber: req.body.phoneNumber,
+        username: req.body.username,
+        subscribed: true,
+        groupID: nreq.data.groupID
       });
-    });
+
+      // attempt to save the user
+      newUser.save(err => {
+        if (err) {
+          console.log(err);
+          return res.json({ success: false, message: err });
+        }
+        console.log("made user");
+        res.json({
+          success: true,
+          message: "Successfully created new user"
+        });
+      });
+    } catch (err) {
+      console.error(err);
+      process.exit(1);
+    }
   }
 };
 
@@ -80,6 +91,10 @@ exports.deleteUser = (req, res) => {
     // make sure that the user exists
     if (count > 0) {
       // remove the user that matches the username
+      User.remove({ username: req.body.username }, (err, bear) => {
+        //if (err) res.send(err);
+       // res.json({ success: true, message: "Successfully unsubscribed" });
+      });
         
       //free up the phone number
         try {
